Backend/requirements.txt
fastapi
uvicorn
mcp
python-dotenv
requests
httpx 
pydantic 
a2a-sdk[http-server]>=0.3.0
langgraph 
google-adk
langchain-mcp-adapters
motor
httpx
mistralai
langchain
opik
psycopg2
fastmcp